requirements.txt
numpy 
pandas
python-dotenv 
mysql-connector-python 
pymysql
scikit-learn 
sqlalchemy



#-e.
